refactor: move `Input` component to app/shared components scope
BREAKING CHANGE: the `Input` component no longer exist in auth/shared scope.
