+++ src/app/screens/auth/shared/components/form/index.ts
@@ -1,3 +1,2 @@
-export * from './input';
 export * from './checkbox';
 export * from './status';
