+++ src/app/shared/components/index.ts
@@ -3,3 +3,4 @@ export * from './backdrop';
 export * from './action-sheet';
 export * from './route';
 export * from './button';
+export * from './form';
